extract useNodeState
